+++ tbot/loader.py
@@ -1,3 +1,19 @@
+# TBot, Embedded Automation Tool
+# Copyright (C) 2018  Harald Seiler
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import sys
 import importlib
 import pathlib
@@ -6,7 +22,7 @@ from termcolor2 import c
 
 
 def list_dir(
-    d: pathlib.Path, recurse: bool = False
+    d: pathlib.Path, recurse: bool = True
 ) -> typing.Generator[pathlib.Path, None, None]:
     for f in d.iterdir():
         if recurse and f.is_dir() and f.name != "__pycache__":