+++ tbot/machine/board/uboot.py
@@ -1,9 +1,26 @@
+# TBot, Embedded Automation Tool
+# Copyright (C) 2018  Harald Seiler
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 import typing
 import shlex
 import tbot
 from tbot import machine
 from tbot.machine import board
 from tbot.machine import linux
+from tbot.machine import channel
 from . import special
 
 B = typing.TypeVar("B", bound=board.Board)
@@ -28,7 +45,10 @@ class UBootMachine(board.BoardMachine[B], machine.InteractiveMachine):
 
     .. py:attribute:: autoboot_prompt: str
 
-        Regular expression of the autoboot prompt
+        Regular expression of the autoboot prompt, set to ``None`` if
+        autoboot is disabled for this board.
+
+        Defaults to ``r"Hit any key to stop autoboot:\s+\d+\s+"``.
 
     .. py:attribute:: autoboot_keys: str = "\\n"
 
@@ -39,7 +59,7 @@ class UBootMachine(board.BoardMachine[B], machine.InteractiveMachine):
         U-Prompt that was configured when building U-Boot
     """
 
-    autoboot_prompt = r"Hit any key to stop autoboot:\s+\d+\s+"
+    autoboot_prompt: typing.Optional[str] = r"Hit any key to stop autoboot:\s+\d+\s+"
     autoboot_keys = "\n"
     prompt = "U-Boot> "
 
@@ -51,7 +71,7 @@ class UBootMachine(board.BoardMachine[B], machine.InteractiveMachine):
     def __init__(self, board: B) -> None:  # noqa: D107
         super().__init__(board)
 
-        self.channel: machine.channel.Channel
+        self.channel: channel.Channel
         if board.channel is not None:
             self.channel = board.channel
         else:
@@ -64,13 +84,16 @@ class UBootMachine(board.BoardMachine[B], machine.InteractiveMachine):
         ) as boot_ev:
             boot_ev.verbosity = tbot.log.Verbosity.STDOUT
             boot_ev.prefix = "   <> "
-            boot_log = self.channel.read_until_prompt(
-                self.autoboot_prompt, regex=True, stream=boot_ev
-            )
+            if self.autoboot_prompt is not None:
+                boot_log = self.channel.read_until_prompt(
+                    self.autoboot_prompt, regex=True, stream=boot_ev
+                )
 
-            boot_ev.data["output"] = boot_log
-        self.channel.send(self.autoboot_keys)
-        self.channel.read_until_prompt(self.prompt)
+                boot_ev.data["output"] = boot_log
+                self.channel.send(self.autoboot_keys)
+                self.channel.read_until_prompt(self.prompt)
+            else:
+                self.channel.read_until_prompt(self.prompt, stream=boot_ev)
 
     def destroy(self) -> None:
         """Destroy this U-Boot machine."""