+++ tbot/tc/selftest/board_machine.py
@@ -1,8 +1,10 @@
+import contextlib
 import typing
 import tbot
 from tbot.machine import channel
 from tbot.machine import linux
 from tbot.machine import board
+from . import machine as mach
 
 
 class TestBoard(board.Board):
@@ -17,9 +19,10 @@ class TestBoard(board.Board):
         self.lh.exec0("rm", self.lh.workdir / "selftest_power")
 
     def connect(self) -> channel.Channel:  # noqa: D102
-        return self.lh.new_channel(
-            linux.Raw(
-                """\
+        if self.has_autoboot:
+            return self.lh.new_channel(
+                linux.Raw(
+                    """\
 bash --norc --noediting; exit
 unset HISTFILE
 PS1='Test-U-Boot> '
@@ -38,8 +41,33 @@ unset HISTFILE
 set +o emacs
 set +o vi
 read -p 'Autoboot: '"""
+                )
             )
-        )
+        else:
+            return self.lh.new_channel(
+                linux.Raw(
+                    """\
+bash --norc --noediting; exit
+unset HISTFILE
+alias version="uname -a"
+function printenv() {
+    if [ $# = 0 ]; then
+        set | grep -E '^U'
+    else
+        set | grep "$1" | sed "s/'//g"
+    fi
+}
+function setenv() { local var="$1"; shift; eval "$var=\\"$*\\""
+}
+PS1=Test-U-Boot'> ' #"""
+                )
+            )
+
+    def __init__(
+        self, lh: linux.LabHost, has_autoboot: bool = True
+    ) -> None:  # noqa: D107
+        self.has_autoboot = has_autoboot
+        super().__init__(lh)
 
 
 class TestBoardUBoot(board.UBootMachine[TestBoard]):
@@ -52,50 +80,75 @@ class TestBoardUBoot(board.UBootMachine[TestBoard]):
 @tbot.testcase
 def selftest_board_uboot(lab: typing.Optional[tbot.selectable.LabHost] = None) -> None:
     """Test if TBot intercepts U-Boot correctly."""
-    with lab or tbot.acquire_lab() as lh:
-        b_: board.Board
+    with contextlib.ExitStack() as cx:
+        lh = cx.enter_context(lab or tbot.acquire_lab())
         try:
-            b_ = tbot.acquire_board(lh)
+            b = cx.enter_context(tbot.acquire_board(lh))
+            ub = cx.enter_context(tbot.acquire_uboot(b))
         except NotImplementedError:
-            b_ = TestBoard(lh)
-        with b_ as b:
-            ub_: board.UBootMachine
-            if isinstance(b, tbot.selectable.Board):
-                ub_ = tbot.acquire_uboot(b)
-            elif isinstance(b, TestBoard):
-                ub_ = TestBoardUBoot(b)
-            with ub_ as ub:
-                ub.exec0("version")
-                env = ub.exec0("printenv").strip().split("\n")
+            b = cx.enter_context(TestBoard(lh))
+            ub = cx.enter_context(TestBoardUBoot(b))
 
-                for line in env[:-1]:
-                    if line != "" and line[0].isalnum():
-                        assert "=" in line, repr(line)
+        ub.exec0("version")
+        env = ub.exec0("printenv").strip().split("\n")
 
-                out = ub.exec0("echo", board.F("0x{}", str(1234))).strip()
-                assert out == "0x1234", repr(out)
+        for line in env[:-1]:
+            if line != "" and line[0].isalnum():
+                assert "=" in line, repr(line)
 
-                from . import machine as mach
+        out = ub.exec0("echo", board.F("0x{}", str(1234))).strip()
+        assert out == "0x1234", repr(out)
 
-                mach.selftest_machine_shell(ub)
+        mach.selftest_machine_shell(ub)
+
+
+@tbot.testcase
+def selftest_board_uboot_noab(
+    lab: typing.Optional[tbot.selectable.LabHost] = None
+) -> None:
+    """Test if TBot intercepts U-Boot correctly without autoboot."""
+
+    class TestBoardUBootNoAB(board.UBootMachine[TestBoard]):
+        """Dummy Board UBoot."""
+
+        autoboot_prompt = None
+        prompt = "Test-U-Boot> "
+
+    with contextlib.ExitStack() as cx:
+        lh = cx.enter_context(lab or tbot.acquire_lab())
+        b = cx.enter_context(TestBoard(lh, has_autoboot=False))
+        ub = cx.enter_context(TestBoardUBootNoAB(b))
+
+        ub.exec0("version")
+        env = ub.exec0("printenv").strip().split("\n")
+
+        for line in env[:-1]:
+            if line != "" and line[0].isalnum():
+                assert "=" in line, repr(line)
+
+        out = ub.exec0("echo", board.F("0x{}", str(1234))).strip()
+        assert out == "0x1234", repr(out)
+
+        mach.selftest_machine_shell(ub)
 
 
 @tbot.testcase
 def selftest_board_linux(lab: typing.Optional[tbot.selectable.LabHost] = None) -> None:
     """Test board's linux."""
-    with lab or tbot.acquire_lab() as lh:
+    with contextlib.ExitStack() as cx:
+        lh = cx.enter_context(lab or tbot.acquire_lab())
+
         try:
-            b_ = tbot.acquire_board(lh)
+            b = cx.enter_context(tbot.acquire_board(lh))
         except NotImplementedError:
             tbot.log.message(
                 tbot.log.c("Skipped").yellow.bold + " because no board available."
             )
             return
-        with b_ as b:
-            with tbot.acquire_linux(b) as lnx:
-                from . import machine as mach
 
-                mach.selftest_machine_shell(lnx)
+        lnx = cx.enter_context(tbot.acquire_linux(b))
+
+        mach.selftest_machine_shell(lnx)
 
 
 @tbot.testcase