+++ tbot/tc/uboot/__init__.py
@@ -1,0 +1,4 @@
+from .build_info import BuildInfo
+from .build import build
+
+__all__ = ("BuildInfo", "build")